Update relayers to work with channel IDs
