+++ relayer/cmd/run/execution/command.go
@@ -2,11 +2,15 @@ package execution
 
 import (
 	"context"
+	"encoding/hex"
 	"log"
 	"os"
 	"os/signal"
+	"reflect"
+	"strings"
 	"syscall"
 
+	"github.com/mitchellh/mapstructure"
 	"github.com/sirupsen/logrus"
 	"github.com/snowfork/snowbridge/relayer/chain/parachain"
 	"github.com/snowfork/snowbridge/relayer/relays/execution"
@@ -50,7 +54,7 @@ func run(_ *cobra.Command, _ []string) error {
 	}
 
 	var config execution.Config
-	err := viper.Unmarshal(&config)
+	err := viper.Unmarshal(&config, viper.DecodeHook(HexHookFunc()))
 	if err != nil {
 		return err
 	}
@@ -100,4 +104,50 @@ func run(_ *cobra.Command, _ []string) error {
 	}
 
 	return nil
+}
+
+func HexHookFunc() mapstructure.DecodeHookFuncType {
+	return func(
+		f reflect.Type,
+		t reflect.Type,
+		data interface{},
+	) (interface{}, error) {
+		// Check that the data is string
+		if f.Kind() != reflect.String {
+			return data, nil
+		}
+
+		// Check that the target type is our custom type
+		if t != reflect.TypeOf(execution.ChannelID{}) {
+			return data, nil
+		}
+
+		foo, err := HexDecodeString(data.(string))
+		if err != nil {
+			return nil, err
+		}
+
+		var out [32]byte
+		copy(out[:], foo)
+
+		// Return the parsed value
+		return execution.ChannelID(out), nil
+	}
+}
+
+// HexDecodeString decodes bytes from a hex string. Contrary to hex.DecodeString, this function does not error if "0x"
+// is prefixed, and adds an extra 0 if the hex string has an odd length.
+func HexDecodeString(s string) ([]byte, error) {
+	s = strings.TrimPrefix(s, "0x")
+
+	if len(s)%2 != 0 {
+		s = "0" + s
+	}
+
+	b, err := hex.DecodeString(s)
+	if err != nil {
+		return nil, err
+	}
+
+	return b, nil
 }
