+++ relayer/relays/execution/config.go
@@ -1,10 +1,6 @@
 package execution
 
 import (
-	"encoding/hex"
-	"errors"
-	"strings"
-
 	"github.com/snowfork/snowbridge/relayer/config"
 )
 
@@ -28,34 +24,3 @@ type SinkConfig struct {
 }
 
 type ChannelID [32]byte
-
-func (m *ChannelID) UnmarshalJSON(bz []byte) error {
-	data, err := HexDecodeString(string(bz))
-	if err != nil {
-		return err
-	}
-	if len(data) != 32 {
-		return errors.New("Invalid Channel ID")
-	}
-
-	copy(m[:], data)
-
-	return nil
-}
-
-// HexDecodeString decodes bytes from a hex string. Contrary to hex.DecodeString, this function does not error if "0x"
-// is prefixed, and adds an extra 0 if the hex string has an odd length.
-func HexDecodeString(s string) ([]byte, error) {
-	s = strings.TrimPrefix(s, "0x")
-
-	if len(s)%2 != 0 {
-		s = "0" + s
-	}
-
-	b, err := hex.DecodeString(s)
-	if err != nil {
-		return nil, err
-	}
-
-	return b, nil
-}
