+++ web/packages/test/scripts/start-relayer.sh
@@ -17,12 +17,12 @@ config_relayer() {
     ' \
         config/beefy-relay.json >$output_dir/beefy-relay.json
 
-    # Configure parachain relay (bridge hub)
+    # Configure parachain relay (primary governance)
     jq \
         --arg k1 "$(address_for GatewayProxy)" \
         --arg k2 "$(address_for BeefyClient)" \
         --arg eth_endpoint_ws $eth_endpoint_ws \
-        --arg channelID $BRIDGE_HUB_PARAID \
+        --arg channelID $PRIMARY_GOVERNANCE_CHANNEL_ID \
         --arg eth_gas_limit $eth_gas_limit \
         '
       .source.contracts.Gateway = $k1
@@ -33,14 +33,32 @@ config_relayer() {
     | .sink.ethereum."gas-limit" = $eth_gas_limit
     | .source."channel-id" = $channelID
     ' \
-        config/parachain-relay.json >$output_dir/parachain-relay-bridge-hub.json
+        config/parachain-relay.json >$output_dir/parachain-relay-bridge-hub-01.json
+
+    # Configure parachain relay (secondary governance)
+    jq \
+        --arg k1 "$(address_for GatewayProxy)" \
+        --arg k2 "$(address_for BeefyClient)" \
+        --arg eth_endpoint_ws $eth_endpoint_ws \
+        --arg channelID $SECONDARY_GOVERNANCE_CHANNEL_ID \
+        --arg eth_gas_limit $eth_gas_limit \
+        '
+      .source.contracts.Gateway = $k1
+    | .source.contracts.BeefyClient = $k2
+    | .sink.contracts.Gateway = $k1
+    | .source.ethereum.endpoint = $eth_endpoint_ws
+    | .sink.ethereum.endpoint = $eth_endpoint_ws
+    | .sink.ethereum."gas-limit" = $eth_gas_limit
+    | .source."channel-id" = $channelID
+    ' \
+        config/parachain-relay.json >$output_dir/parachain-relay-bridge-hub-02.json
 
     # Configure parachain relay (asset hub)
     jq \
         --arg k1 "$(address_for GatewayProxy)" \
         --arg k2 "$(address_for BeefyClient)" \
         --arg eth_endpoint_ws $eth_endpoint_ws \
-        --arg channelID $ASSET_HUB_PARAID \
+        --arg channelID $ASSET_HUB_CHANNEL_ID \
         --arg eth_gas_limit $eth_gas_limit \
         '
       .source.contracts.Gateway = $k1
@@ -58,7 +76,7 @@ config_relayer() {
         --arg k1 "$(address_for GatewayProxy)" \
         --arg k2 "$(address_for BeefyClient)" \
         --arg eth_endpoint_ws $eth_endpoint_ws \
-        --arg channelID $TEMPLATE_PARA_ID \
+        --arg channelID $TEMPLATE_CHANNEL_ID \
         --arg eth_gas_limit $eth_gas_limit \
         '
       .source.contracts.Gateway = $k1
@@ -85,7 +103,7 @@ config_relayer() {
     jq \
         --arg eth_endpoint_ws $eth_endpoint_ws \
         --arg k1 "$(address_for GatewayProxy)" \
-        --arg channelID $ASSET_HUB_PARAID \
+        --arg channelID $ASSET_HUB_CHANNEL_ID \
         '
       .source.ethereum.endpoint = $eth_endpoint_ws
     | .source.contracts.Gateway = $k1
@@ -97,7 +115,7 @@ config_relayer() {
     jq \
         --arg eth_endpoint_ws $eth_endpoint_ws \
         --arg k1 "$(address_for GatewayProxy)" \
-        --arg channelID $TEMPLATE_PARA_ID \
+        --arg channelID $TEMPLATE_CHANNEL_ID \
         '
               .source.ethereum.endpoint = $eth_endpoint_ws
             | .source.contracts.Gateway = $k1
@@ -121,15 +139,28 @@ start_relayer() {
         done
     ) &
 
-    # Launch parachain relay for bridgehub
+    # Launch parachain relay for bridgehub (primary governance)
     (
-        : >"$output_dir"/parachain-relay-bridge-hub.log
+        : >"$output_dir"/parachain-relay-bridge-hub-01.log
         while :; do
-            echo "Starting parachain-relay (bridgehub) at $(date)"
+            echo "Starting parachain-relay (primary governance) at $(date)"
             "${relay_bin}" run parachain \
-                --config "$output_dir/parachain-relay-bridge-hub.json" \
+                --config "$output_dir/parachain-relay-bridge-hub-01.json" \
                 --ethereum.private-key $parachain_relay_eth_key \
-                >>"$output_dir"/parachain-relay-bridge-hub.log 2>&1 || true
+                >>"$output_dir"/parachain-relay-bridge-hub-01.log 2>&1 || true
+            sleep 20
+        done
+    ) &
+
+    # Launch parachain relay for bridgehub (secondary governance)
+    (
+        : >"$output_dir"/parachain-relay-bridge-hub-02.log
+        while :; do
+            echo "Starting parachain-relay (secondary governance) at $(date)"
+            "${relay_bin}" run parachain \
+                --config "$output_dir/parachain-relay-bridge-hub-02.json" \
+                --ethereum.private-key $parachain_relay_eth_key \
+                >>"$output_dir"/parachain-relay-bridge-hub-02.log 2>&1 || true
             sleep 20
         done
     ) &
@@ -181,7 +212,7 @@ start_relayer() {
             "${relay_bin}" run execution \
                 --config $output_dir/execution-relay-asset-hub.json \
                 --substrate.private-key "//ExecutionRelay" \
-                >>"$output_dir"/execution-relay.log 2>&1 || true
+                >>"$output_dir"/execution-relay-asset-hub.log 2>&1 || true
             sleep 20
         done
     ) &
